feat: application/v1beta1
Builds upon #25642. In order to enable CEL validations and to make this
a non-breaking change, introduce `Application/v1beta1`. This means that
users would need to opt-in to the new functionality.

Internally v1beta1 will be stored as v1alpha1, meaning we don't need to
run conversion webhooks, and it also means that there are no changes to
the fields in v1beta1.

However, we use CEL to deprecate the usage of `source` (only `sources`
can be used with v1beta1).

CEL doesn't really like the notion of setting validation rules on arrays
that are potentially unbounded (even though in practice they are really
not).

So to make CEL happy, we set a limit on `sources` (20, but can be
reduced), as well as on ignoreDifferences (100, which can also be
reduced) and `syncOptions` (20)

Other notable changes:

Required fields:
1. `sources` required (not `source`)
2. All sources must have `repoURL`
3. All sources must have `targetRevision`
4. `project` required

Destination rules:
5. Destination must have `server` OR `name`
6. Destination can't have both `server` AND `name`

Source mutual exclusions:
7. Sources can't have both `chart` and `path`
8. Ref sources can't have `path` or `chart` (`ref` sources are pure references)

Numeric constraints:
9. Retry limit >= 0
10. Backoff factor >= 1
11. RevisionHistoryLimit >= 0

IgnoreDifferences:
12. Entries must have a kind

SyncOptions conflicts (8 total):
13. PrunePropagationPolicy enum validation (must be one of: `background`,
`foreground`, `orphan`)
14. `Replace=true/false` conflict
15. `ServerSideApply=true/false` conflict
16. `ApplyOutOfSyncOnly=true/false` conflict
17. `ClientSideApplyMigration=true/false` conflict
18. `Prune=false` vs `Prune=confirm` conflict
19. `Delete=false` vs `Delete=confirm` conflict
20. `PruneLast=true` with `Prune=false` conflict

SyncOptions type:
21. Enum validation (22 valid options)

Signed-off-by: Blake Pettersson <blake.pettersson@gmail.com>